requirements.txt
streamlit
python-binance
python-dotenv